+++ src/automated-polymorpher.ts
@@ -94,3 +94,12 @@ Hooks.once('ready', () => {
 });
 
 // Add any additional hooks if necessary
+
+Hooks.once('libChangelogsReady', function () {
+  //@ts-ignore
+  libChangelogs.register(
+    AUTOMATED_POLYMORPHER_MODULE_NAME,
+    '- Add [CHANGELOGS & CONFLICTS](https://github.com/theripper93/libChangelogs) hooks for better management of the conflicts',
+    'minor',
+  );
+});