+++ changelog.md
@@ -1,1 +1,5 @@
-# CHANGELOG
+## 1.0.8
+
+- Add [CHANGELOGS & CONFLICTS](https://github.com/theripper93/libChangelogs) hooks for better management of the conflicts
+- Colored the header sheet button icon and text
+- Add module setting to hide the header button sheet label (active by default)